+++ packaging/kde/package-kde-store-archive.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-VERSION="1.2.1"
+VERSION="1.2.2"
 ARCHIVE_NAME="ltmnight-sddm-theme-v${VERSION}.tar.gz"
 
 # Find the repo root